Standardize namespace to Spree::GraphQL (#7, 2)
Now all classes are defined under Spree::GraphQL, except for
Spree::GraphqlController which is named this way due to autoloader.
