+++ lib/solidus_graphql_api/engine.rb
@@ -1,20 +1,22 @@
-module SolidusGraphqlApi
-  class Engine < Rails::Engine
-    require 'spree/core'
-    isolate_namespace Spree
-    engine_name 'solidus_graphql_api'
+module Spree
+  module GraphQL
+    class Engine < Rails::Engine
+      require 'spree/core'
+      isolate_namespace Spree
+      engine_name 'solidus_graphql'
 
-    # use rspec for tests
-    config.generators do |g|
-      g.test_framework :rspec
-    end
-
-    def self.activate
-      Dir.glob(File.join(File.dirname(__FILE__), '../../app/**/*_decorator*.rb')) do |c|
-        Rails.configuration.cache_classes ? require(c) : load(c)
+      # use rspec for tests
+      config.generators do |g|
+        g.test_framework :rspec
       end
-    end
 
-    config.to_prepare(&method(:activate).to_proc)
+      def self.activate
+        Dir.glob(File.join(File.dirname(__FILE__), '../../app/**/*_decorator*.rb')) do |c|
+          Rails.configuration.cache_classes ? require(c) : load(c)
+        end
+      end
+
+      config.to_prepare(&method(:activate).to_proc)
+    end
   end
 end
