+++ lib/solidus_graphql_api/graphql/mutations/base_mutation.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     module Mutations
       class BaseMutation < ::GraphQL::Schema::RelayClassicMutation
