+++ lib/solidus_graphql_api/graphql/resolvers/option_type_resolver.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class OptionTypeResolver
       class All
