+++ lib/solidus_graphql_api/graphql/resolvers/order_resolver.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
 
     class OrderResolver
