+++ lib/solidus_graphql_api/graphql/resolvers/payment_resolver.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class PaymentsResolver
       class ByOrderId
