+++ lib/solidus_graphql_api/graphql/resolvers/product_resolver.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class ProductResolver
       class All
