+++ lib/solidus_graphql_api/graphql/resolvers/shipment_method_resolver.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class ShipmentMethodResolver
       class All
