+++ lib/solidus_graphql_api/graphql/schema.rb
@@ -1,4 +1,8 @@
-class SolidusSchema < GraphQL::Schema
-  mutation(::Solidus::GraphQL::MutationType)
-  query(::Solidus::GraphQL::QueryType)
+module Spree
+  module GraphQL
+    class Schema < ::GraphQL::Schema
+      mutation(::Spree::GraphQL::MutationType)
+      query(::Spree::GraphQL::QueryType)
+    end
+  end
 end
