+++ lib/solidus_graphql_api/graphql/types/address_type.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class AddressType < ::GraphQL::Schema::Object
       graphql_name "Address"
