+++ lib/solidus_graphql_api/graphql/types/attachment_type.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class Attachment < ::GraphQL::Schema::Object
     graphql_name "Attachment"
