+++ lib/solidus_graphql_api/graphql/types/base_input_object.rb
@@ -1,2 +1,2 @@
-class Types::BaseInputObject < GraphQL::Schema::InputObject
+class Types::BaseInputObject < ::GraphQL::Schema::InputObject
 end
