+++ lib/solidus_graphql_api/graphql/types/base_interface.rb
@@ -1,3 +1,3 @@
 module Types::BaseInterface
-  include GraphQL::Schema::Interface
+  include ::GraphQL::Schema::Interface
 end
