+++ lib/solidus_graphql_api/graphql/types/image_type.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     require_relative './attachment_type'
     class ImageType < ::Types::BaseObject
