+++ lib/solidus_graphql_api/graphql/types/line_item_type.rb
@@ -1,5 +1,5 @@
 require_relative 'variant_type'
-module Solidus
+module Spree
   module GraphQL
     class LineItemType < Types::BaseObject
       graphql_name "LineItem"
