+++ lib/solidus_graphql_api/graphql/types/option_type.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class OptionType < ::Types::BaseObject
       graphql_name "OptionType"
