+++ lib/solidus_graphql_api/graphql/types/option_value_type.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class OptionValueType  < ::Types::BaseObject
       graphql_name "OptionValue"
