+++ lib/solidus_graphql_api/graphql/types/payment_type.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class PaymentType < Types::BaseObject
       graphql_name "Payment"
