+++ lib/solidus_graphql_api/graphql/types/query_type.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
     class QueryType < ::Types::BaseObject
