+++ lib/solidus_graphql_api/graphql/types/variant_type.rb
@@ -1,4 +1,4 @@
-module Solidus
+module Spree
   module GraphQL
    class VariantType < ::GraphQL::Schema::Object
