+++ spec/graphql/mutations/order_mutation_spec.rb
@@ -21,7 +21,7 @@ RSpec.describe "Order mutation" do
        }
       }
       GRAPHQL
-      response = SolidusSchema.execute(query_str)
+      response = Spree::GraphQL::Schema.execute(query_str)
       expect(response["data"]["createOrder"]["orderToken"]).not_to be nil
     end
 
@@ -38,7 +38,7 @@ RSpec.describe "Order mutation" do
        }
       }
       GRAPHQL
-      response = SolidusSchema.execute(query_str)
+      response = Spree::GraphQL::Schema.execute(query_str)
       orderToken = response["data"]["createOrder"]["orderToken"]
       add_item_query_str = <<-GRAPHQL
       mutation{
@@ -54,7 +54,7 @@ RSpec.describe "Order mutation" do
       }
      }
       GRAPHQL
-      response = SolidusSchema.execute(add_item_query_str)
+      response = Spree::GraphQL::Schema.execute(add_item_query_str)
       puts "Info #{response.to_json}"
       quantity=response["data"]["addOrUpdateItem"]["lineItem"]["quantity"]
       expect(quantity.to_i).to be 2
